Re-organize example code
Move from `usage` to `examples`
Remove `express` examples, since the express middleware has been broken out into it's own package.
