+++ examples/client/SampleClient.js
@@ -1,5 +1,5 @@
 //
-//  /usage/client/SampleClient.js
+//  /examples/client/SampleClient.js
 //  Example of a SimpleHMACAuth Client subclass for a service
 //  This client may be used with either callbacks or promises
 //  Created by Jesse T Youngblood on 11/23/18 at 19:11pm
