+++ examples/client/additional/BearClient.js
