+++ examples/client/additional/ChickenClient.js
