+++ examples/client/additional/bear_usage.js
