+++ examples/client/direct_promises_async.js
@@ -1,7 +1,7 @@
 //
 //  Simple HMAC Auth
 //  Direct, non-subclassed usage of the Client class
-//  /usage/client/direct_promises_async.js
+//  /examples/client/direct_promises_async.js
 //  Created by Jesse T Youngblood on 8/13/18 at 18:50
 //
 
@@ -17,7 +17,6 @@ const client = new SimpleHMACAuth.Client('API_KEY', 'SECRET', {
   ssl: false,
   verbose: true
 });
-
 
 (async () => {
