+++ examples/client/sample_callbacks.js
@@ -1,7 +1,7 @@
 //
 //  Simple HMAC Auth
-//  /usage/client/sample_promises.js
-//  Created by Jesse T Youngblood on 11/23/18 at 19:14
+//  /examples/client/sample_callbacks.js
+//  Created by Jesse T Youngblood on 3/23/16 at 10:42pm
 //
 
 /* eslint no-console: off */
@@ -22,11 +22,12 @@ const query = {
   array: [ 1, 2, 3 ]
 };
 
-client.query(query).then(results => {
+client.query(query, (error, results) => {
+
+  if (error) {
+    console.error(`Received error:`, error);
+    return;
+  }
 
   console.log(results);
-
-}).catch(error => {
-
-  console.log(`Received error:`, error);
 });
