--- usage/server/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "simple-hmac-auth-usage-server-examples",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {},
-  "dependencies": {
-    "express": "^4.16.4"
-  }
-}
